+++ package.json
@@ -16,7 +16,7 @@
     "react-scripts": "4.0.3",
     "redux": "^4.2.0",
     "redux-thunk": "^2.4.1",
-    "web-vitals": "^1.0.1"
+    "uuid": "^8.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
