--- src/actions/actions.js
@@ -1,10 +1,0 @@
-
-const SET_ORDER = 'SET_ORDER';
-const SET_PRICE = 'SET_PRICE';
-const RESET     = 'RESET';
-
-export {
-    SET_ORDER,
-    SET_PRICE,
-    RESET
-}
