+++ src/services/actions/ingredients.js
@@ -3,8 +3,6 @@ import { getIngredientsRequest } from "../../utils/request";
 export const GET_INGREDIENTS_REQUEST = 'GET_INGREDIENTS_REQUEST';
 export const GET_INGREDIENTS_SUCCESS = 'GET_INGREDIENTS_SUCCESS';
 export const GET_INGREDIENTS_FAILED  = 'GET_INGREDIENTS_FAILED';
-export const SHOW_INGREDIENT_DETAILS = 'SHOW_INGREDIENT_DETAILS';
-export const HIDE_INGREDIENT_DETAILS = 'HIDE_INGREDIENT_DETAILS';
 
 export const SET_ACTIVE_TAB = 'SET_ACTIVE_TAB';
 
@@ -18,10 +16,5 @@ export const getIngredients = () => {
         })
     }
 };
-
-export const showIngredientDetails = id => dispatch => dispatch({type: SHOW_INGREDIENT_DETAILS, id});
-
-export const hideIngredientDetails = () => dispatch => dispatch({type: HIDE_INGREDIENT_DETAILS});
-
 export const setActiveTab = cat => dispatch => dispatch({type: SET_ACTIVE_TAB, cat});
