+++ src/services/actions/orderDetails.js
@@ -1,0 +1,18 @@
+import { getOrderRequest } from "../../utils/request";
+
+export const GET_ORDER_REQUEST = 'GET_ORDER_REQUEST';
+export const GET_ORDER_SUCCESS = 'GET_ORDER_SUCCESS';
+export const GET_ORDER_FAILED  = 'GET_ORDER_FAILED';
+export const CLOSE_ORDER = 'CLOSE_ORDER';
+
+export const getOrder = ingredients => {
+    return (dispatch) => {
+        dispatch({type: GET_ORDER_REQUEST});
+        getOrderRequest(ingredients).then(({ name, number }) => {
+            dispatch({type: GET_ORDER_SUCCESS, name, number});
+        }).catch(() => {
+            dispatch({type: GET_ORDER_FAILED});
+        })
+    }
+};
+export const closeOrder = () => dispatch => dispatch({type: CLOSE_ORDER})
