+++ src/services/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from "redux";
-import { ingredientsReducer, ingredientDetailsReducer } from "./ingredients";
-import { constructorReducer, orderDetailsReducer } from "./constructor";
+import ingredientsReducer from "./ingredients";
+import ingredientDetailsReducer from "./ingredientDetails";
+import constructorReducer from "./constructor";
+import orderDetailsReducer from "./orderDetails";
 
 export const rootReducer = combineReducers({
     ingredients: ingredientsReducer,
